// web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        UserDto: {
            /** Format: date-time */
            approved_at?: string | null;
            /** Format: uuid */
            approved_by?: string | null;
            /** Format: date-time */
            created_at: string;
            display_name: string;
            email?: string | null;
            email_verified: boolean;
            /** Format: uuid */
            id: string;
            /** Format: date-time */
            last_seen_at?: string | null;
            role: components["schemas"]["RoleDto"];
            status: components["schemas"]["StatusDto"];
            /** Format: date-time */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}